update error message

+++ controllers/users.js
@@ -4,7 +4,23 @@ const User = require('../helpers/users');
 const Post = require('../helpers/posts');
 const PostsController = require('./posts');
 
-exports.signup = (req, res) => {
+exports.signup = async (req, res) => {
+  if (
+    !req.body.firstName ||
+    !req.body.lastName ||
+    !this.validateEmail(req.body.email) ||
+    !this.validatePassword(req.body.password)
+  ) {
+    return res
+      .status(400)
+      .json({ error: 'Sent data does not match requirements' });
+  }
+  const existingUser = await this.findUserByEmail(req.body.email);
+  if (existingUser) {
+    return res
+      .status(409)
+      .json({ error: 'A user already exist with given email address' });
+  }
   bcrypt
     .hash(req.body.password, 10)
     .then(async (hash) => {
@@ -24,18 +40,25 @@ exports.signup = (req, res) => {
             }),
           })
         )
-        .catch((error) => res.status(400).json({ error }));
+        .catch((error) => res.status(500).json({ error }));
     })
     .catch((error) => res.status(500).json({ error }));
 };
 
-exports.login = (req, res) => {
+exports.login = async (req, res) => {
+  if (
+    !this.validateEmail(req.body.email) ||
+    !this.validatePassword(req.body.password)
+  ) {
+    return res
+      .status(400)
+      .json({ error: 'Sent data does not match requirements' });
+  }
   User.findOne({ where: { email: req.body.email } })
     .then((user) => {
       if (!user) {
         return res.status(404).json({ error: 'Utilisateur non trouvé !' });
       }
-
       bcrypt
         .compare(req.body.password, user.password)
         .then((valid) => {
@@ -66,17 +89,8 @@ exports.getAllUsers = (req, res) => {
     });
 };
 
-exports.getOneUser = (req, res) => {
+exports.getOneUser = async (req, res) => {
   User.findOne({
-    attributes: [
-      'id',
-      'firstName',
-      'lastName',
-      'email',
-      'password',
-      'createdAt',
-      'updatedAt',
-    ],
     where: { id: req.params.userId },
   })
     .then((user) => {
@@ -92,18 +106,47 @@ exports.getOneUser = (req, res) => {
 };
 
 exports.modifyUser = async (req, res) => {
-  User.update({ ...req.body }, { where: { id: req.params.userId } })
+  if (
+    !req.body.firstName ||
+    !req.body.lastName ||
+    !this.validateEmail(req.body.email) ||
+    !this.validatePassword(req.body.password)
+  ) {
+    return res
+      .status(400)
+      .json({ error: 'Sent data does not match requirements' });
+  }
+  const existingUser = await this.findUserByEmail(req.body.email);
+  if (existingUser) {
+    return res
+      .status(409)
+      .json({ error: 'A user already exist with given email address' });
+  }
+  User.findOne({ where: { id: req.params.userId } })
     .then(() => {
-      res.status(201).json({ message: 'Profil modifié !' });
+      User.update({ ...req.body }, { where: { id: req.params.userId } })
+        .then(() => {
+          res.status(201).json({ message: 'Profil modifié !' });
+        })
+        .catch((error) =>
+          res
+            .status(400)
+            .json({ error: 'Sent data does not match requirements' })
+        );
     })
-    .catch((error) => res.status(400).json({ error }));
+    .catch((error) => res.status(404).json({ error: 'no user match this id' }));
 };
 
 exports.deleteUser = async (req, res) => {
-  await PostsController.deletePostsByUserId(req.params.userId);
-  User.destroy({ where: { id: req.params.userId } })
-    .then(() => res.status(200).json({ message: 'Utilisateur supprimé' }))
-    .catch((error) => res.status(400).json({ error }));
+  User.findOne({ where: { id: req.params.userId } }).then(async (user) => {
+    if (!user) {
+      return res.status(404).json({ error: 'no user match this id' });
+    }
+    await PostsController.deletePostsByUserId(req.params.userId);
+    User.destroy({ where: { id: req.params.userId } })
+      .then(() => res.status(200).json({ message: 'Utilisateur supprimé' }))
+      .catch((error) => res.status(400).json({ error }));
+  });
 };
 
 exports.getPostsWithUserId = (req, res) => {
@@ -126,5 +169,32 @@ exports.modifyUserPassword = (req, res) => {
     .then(() => {
       res.status(200).json({ message: 'Mot de Passe modifié ' });
     })
-    .catch((error) => res.status(400).json({ error }));
+    .catch((error) => res.status(404).json({ error: 'no user match this id' }));
+};
+
+exports.findUserByEmail = (email) =>
+  new Promise((resolve, reject) => {
+    User.findOne({ where: { email } })
+      .then((user) => {
+        resolve(user);
+      })
+      .catch((error) => {
+        reject(error);
+      });
+  });
+
+exports.validateEmail = (email) => {
+  if (!email) {
+    return false;
+  }
+  const regex = new RegExp('[a-zA-Z0-9._-]+@[a-zA-Z0-9._-]+.[a-zA-Z.]{2,25}');
+  return regex.test(email);
+};
+
+exports.validatePassword = (password) => {
+  if (!password) {
+    return false;
+  }
+  const regex = new RegExp('.{8,}');
+  return regex.test(password);
 };
